refactor: ui 패키지 도메인별로 분리
