+++ src/main/java/kitchenpos/ui/menugroup/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.ui.menugroup;
 
 import kitchenpos.dto.common.ErrorResponse;
 import kitchenpos.exception.common.BadRequestException;
