+++ src/main/java/kitchenpos/ui/order/OrderRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.ui.order;
 
 import kitchenpos.application.order.OrderService;
 import kitchenpos.dto.order.ChangeOrderStatusRequest;
