+++ src/main/java/kitchenpos/ui/product/ProductRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.ui.product;
 
 import kitchenpos.application.product.ProductService;
 import kitchenpos.dto.product.CreateProductRequest;
